+++ tests/Pest.php
@@ -1,5 +1,7 @@
 <?php
 
+use Flexpik\FilamentStudio\Tests\SpatieTestCase;
 use Flexpik\FilamentStudio\Tests\TestCase;
 
 uses(TestCase::class)->in('Unit', 'Feature');
+uses(SpatieTestCase::class)->in('Integration');